# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking AI - API Documentation",
        "description": "API REST para el sistema de gestión de parqueaderos multi-sucursal. Permite administrar sucursales, usuarios, vehículos, tarifas, descuentos, impuestos y operaciones de entrada/salida de vehículos.",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@parking-ai.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "http://localhost:8001/api",
            "description": "API Base URL"
        }
    ],
    "paths": {
        "/api/admin/branches": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Listar sucursales",
                "description": "Listar sucursales\n\nObtiene todas las sucursales del sistema con información del usuario responsable",
                "operationId": "d73087431bca17088a52862685dadc97",
                "responses": {
                    "200": {
                        "description": "Listado exitoso"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Crear sucursal",
                "description": "Crear sucursal",
                "operationId": "fdd7a7a623c5fbb562cb4a3c98052c49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "total_spaces",
                                    "available_spaces"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Sede Norte"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Calle 123 #45-67"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "3001234567"
                                    },
                                    "total_spaces": {
                                        "type": "integer",
                                        "minimum": 1,
                                        "example": 100
                                    },
                                    "available_spaces": {
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 100
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creada exitosamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/branches/{id}": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Obtener sucursal",
                "description": "Obtener sucursal por ID",
                "operationId": "794cd66696e25d36f9d9c75f6204fda9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrada exitosamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Actualizar sucursal",
                "description": "Actualizar sucursal",
                "operationId": "32f926afc84e1b9d862be578f9e006e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "total_spaces",
                                    "available_spaces"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20
                                    },
                                    "total_spaces": {
                                        "type": "integer",
                                        "minimum": 1
                                    },
                                    "available_spaces": {
                                        "type": "integer",
                                        "minimum": 0
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizada exitosamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Eliminar sucursal",
                "description": "Eliminar sucursal",
                "operationId": "363496886a73200454e7d4e2b734c735",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminada exitosamente"
                    },
                    "404": {
                        "description": "No encontrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Laravel Sanctum. Obtén el token usando el endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}